// src/app/module/auth/auth.interface.ts
import { Role } from "../../../generated/prisma/browser"

export interface ILoginUserPayload {
    email: string
    password: string
}

export interface IRegisterPatientPayload {
    name: string
    email: string
    password: string
}

export interface IRequestUser {
    userId: string
    email: string
    name: string
    role: Role
}

export interface IGoogleLoginPayload {
    idToken: string;
}